AnalysisServer/requirements.txt
elasticsearch==8.17.2
fastapi
yara-python
uvicorn
pefile
requests
langchain==0.3.1
langchain-google-genai==2.1.1
langchain-core==0.3.56
langchain-community==0.3.1
langchain-text-splitters==0.3.0
jinja2
starlette
pytz
websockets